Add random-object.js module for generating random questions
