+++ assets/js/game.js
@@ -5,49 +5,15 @@ It uses the data from the db.js file to populate the game page.
 It also contains the functions that control the game play.
 */
 
-import { setDifficulty, getAttempts } from "./game-storage.js";
-import { DATA } from "./db.js";
+import { getAttempts, getCurrentQuestion } from "./game-storage.js";
 
 runGame();
 
 function runGame() {
   displayAttempts(getAttempts());
-  displayQuestion(getRandomQuestion())
+  displayQuestion(getCurrentQuestion());
 }
 
-/**
- * Return a random object from an array.
- * @param {array} array - The array to get a random object from.
- * @returns {object} - A random object from the array.
- */
-function getRandomObject(array) {
-  return array[Math.floor(Math.random() * array.length)];
-}
-
-/**
- * @returns {object} - A random topic object from the DATA array.
- */
-function getRandomTopicObj() {
-  const randomTopicObj = getRandomObject(DATA);
-  return randomTopicObj;
-}
-
-/**
- * Get a random question from a random topic object.
- * @returns {object} - A random question object with the topic name, question, answers and correct answer.
- * E.g: {topic: "Christmas Trivia", question: "What is?", answers: ["a", "b", "c", "d"], correctAnswer: "a"}
- */
-function getRandomQuestion() {
-  const topicObj = getRandomTopicObj();
-  const randomQuestion = getRandomObject(topicObj.questions);
-  const result = {
-    topic: topicObj.topic,
-    question: randomQuestion.question,
-    answers: randomQuestion.answers,
-    correctAnswer: randomQuestion.correctAnswer,
-  };
-  return result;
-}
 
 function displayQuestion(randomQuestionObj) {
   const topic = document.getElementById("topic");
