+++ game.html
@@ -118,6 +118,7 @@
   <script src="assets/js/script.js"></script>
   <script type="module" src="assets/js/db.js"></script>
   <script type="module" src="assets/js/game-storage.js"></script>
+  <script type="module" src="assets/js/random-object.js"></script>
   <script type="module" src="assets/js/game.js"></script>
 </body>
